pyproject.toml
[project]
name = "youtube-summarizer"
version = "1.0.0"
description = "YouTube video summarizer with transcription and AI-powered summaries"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "yt-dlp>=2024.12.23",
    "openai>=1.58.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.uv]
dev-dependencies = []
